+++ scr/opengl/Renderer.h
@@ -23,9 +23,9 @@ public:
     GLFWwindow* getWindow();
 
 
-    void renderTriangle(glm::vec2 p1, glm::vec2 p2, glm::vec2 p3, glm::vec4 color, int layer);
-    void renderQuad(glm::vec2 p1, glm::vec2 p2, glm::vec2 p3, glm::vec2 p4, glm::vec4 color, int layer);
-    void renderPolygon(std::vector<glm::vec2>& points, glm::vec4 color, int layer);
+    void renderTriangle(glm::vec2 p1, glm::vec2 p2, glm::vec2 p3, glm::vec4 color, float layer);
+    void renderQuad(glm::vec2 p1, glm::vec2 p2, glm::vec2 p3, glm::vec2 p4, glm::vec4 color, float layer);
+    void renderPolygon(std::vector<glm::vec2>& points, glm::vec4 color, float layer);
 
     template <class T>
     void autoBatch(std::vector<IVertex*> vertices);
